extensions/api/auth-spi/src/main/java/org/eclipse/dataspaceconnector/api/auth/AuthorizationRequestFilter.java
/*
 *  Copyright (c) 2020 - 2022 Microsoft Corporation
 *
 *  This program and the accompanying materials are made available under the
 *  terms of the Apache License, Version 2.0 which is available at
 *  https://www.apache.org/licenses/LICENSE-2.0
 *
 *  SPDX-License-Identifier: Apache-2.0
 *
 *  Contributors:
 *       Microsoft Corporation - initial API and implementation
 *
 */

package org.eclipse.dataspaceconnector.api.auth;

import jakarta.ws.rs.container.ContainerRequestContext;
import jakarta.ws.rs.container.ContainerRequestFilter;
import jakarta.ws.rs.container.PreMatching;
import jakarta.ws.rs.ext.Provider;
import org.eclipse.dataspaceconnector.api.exception.NotAuthorizedException;

import java.util.Map;
import java.util.stream.Collectors;

@Provider
@PreMatching
public class AuthorizationRequestFilter implements ContainerRequestFilter {
    private final AuthorizationService authorizationService;

    public AuthorizationRequestFilter(AuthorizationService authorizationService) {
        this.authorizationService = authorizationService;
    }

    @Override
    public void filter(ContainerRequestContext requestContext) {
        var headers = requestContext.getHeaders();

        var isAuthorized = authorizationService.isAuthorized(headers.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));

        if (!isAuthorized) {
            throw new NotAuthorizedException();
        }
    }
}
